\name{print}
\docType{methods}

\alias{print}
\alias{print-methods}
\alias{print,Performance-method}
\alias{print,Person-method}
\alias{print,Domain-method}
\alias{print,DomainManager-method}
\alias{print,HumanResourceManager-method}
\alias{print,Visualiser-method}

\title{ Print an Object }

\description{
	Print a description of the MPIA object.
}


\usage{
   print(x)
}

\arguments{
  \item{x}{A \code{Person} object or a \code{Performance} object. }
}

\section{Methods}{
\describe{
	Outputs a short description of the object to the console.
}}

\author{Fridolin Wild <wild@brookes.ac.uk>}
\references{Fridolin Wild (2016): Learning Analytics in R with SNA, LSA, and MPIA, Springer:Berlin.}

\keyword{mpia}
\keyword{print}
