\name{show}
\docType{methods}

\alias{show}
\alias{show-methods}
\alias{show,Performance-method}
\alias{show,Person-method}
\alias{show,Domain-method}
\alias{show,DomainManager-method}
\alias{show,HumanResourceManager-method}
\alias{show,Visualiser-method}

\title{ Show an Object }

\description{

	Display the object by printing it.

}

\usage{
   show(object)
}

\arguments{
  \item{object}{A \code{Person} object or a \code{Performance} object. }
}

\section{Methods}{
\describe{
	Outputs a short description of the object to the console.
}}

\author{Fridolin Wild <fridolin.wild@open.ac.uk>}
\references{Fridolin Wild (2013): Meaningful Purposive Interaction Analysis.}

\keyword{mpia}
\keyword{print}
